requirements.txt
matplotlib==3.6.2
numpy==1.23.4
Pillow==9.5.0
PyYAML==6.0
scipy==1.10.1
seaborn==0.12.2
setuptools==65.4.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.26.1
